Lower minimum required coverage
Most of the code that exists right now is still boiler plate, or is
completely automatically generated. Because of this, the amount of code
that I feel needs to be unit tested is still very low. I don't feel the
need to write tests over anything that's generated by Lombok, and I
don't feel the need to write tests over the boot strap code to start a
Dropwizard application.