Fix some scaling issues with the sudoku board
This is a first attempt to create a better scaling method for the sudoku
board that does not use pure CSS to generate the dimensions for the divs
that contain the board itself. Instead this is generated through JS.
This still needs some fixing (cells do not scale _quite_ properly yet).