Launch backend in EC2 instance
Our terraform configuration now has an EC2 instance that loads, which
has a custom AMI with docker installed. Upon boot up, the AMI will login
to ECR which is hosting our backend docker image, and then will run the
docker image as a docker daemon.

On top of that, the terraform config creates a domain,
backend.sudokurace.io, which will always be set to the IP address of the
EC2 instance, and all traffic gets forwarded to that instance, so we can
always send requests to backend.sudokurace.io instead of having to worry
about the IP address of the instance itself.