Use ELB with SSL termination
Now we're using a classic load balancer with an auto-scaling group with
a single instance. Right now we're putting all game state in the RAM of
the local instance, so we can't actually scale to more than one
instance. We'll fix that in the near future.

Until then, we'll run with a single instance. At the same time, now that
there's a load balancer in front of our auto scaling group, we can
configure the load balancer to have the SSL termination. I've disabled
http://backend.sudokurace.io, and now all traffic must be encrypted at
https://backend.sudokurace.io.