Taint ASG for backend on each CircleCI deploy
If we don't mark the ASG as tainted, then when the Terraform plan gets
applied, the ASG itself appears to not have changed, even though there
should be a new Docker image to run on the EC2 instances. Because of
this, we just manually mark the ASG as needing to be replaced.

Additionally, I added the `create_before_destroy` attribute to ASG and
related resources, to make sure that a new instance is created before
the previous instance is brought down. This is to make sure that there's
no down time between deployments.