Fix bad code practices in InputCell
The InputCell class now has a method that handles keypress events
seperately from chekcing the validity of an input. The onKeyPress event
is now called through handleKeyPress which in turn calls the static
isNumber method that uses a regex to test whether the input contains a
number. Because only a single input (ie character) can be given via the
maxLength="1" property of the input, if it contains a number, it is a
single number. The tests have been updated to reflect this property.