Use latest tag
The problem was that I was never actually pushing the Docker image that
was tagged to be latest. I thought what was going on, was that since the
image had a tag for the $CIRCLE_SHA1, that it would also push to the
latest tag, because they were technically the same image.