Replace isNumber method with more robust isValidInput
The isNumber method has now been replaced with a much more stringent
test of isValidInput. The isValidInput method now checks to see whether
the given input string is longer than a single character and omits the
possibility of of the input of being 0, which is an invalid move in
sudoku.

The handleKeyPress method has also been updated as well as the
.eslintrc.yml file due to the fact that using this.event would appease
eslint but cause a crash on the frontend (even after successfull
compilation) due to this.event not being defined. handleKeyPress should
not be a static method, and should be invoked by each instatiation of
each InputCell on the board. The best course of action was to add it to
the excepted methods in the .eslintrc.yml file.