Specify the CIRCLE_SHA1 for the docker image
This is a combination of fixes to make the backend work again.

Firstly, the AMI that was being used, was actually snapshotted with a
java docker image running, by default. I've fixed that, and created a
new AMI.

Second, the latest tag was being unreliable, so I've removed any need
for it, by passing in the correct CIRCLE_SHA1 directly into the
terraform configuration, so that the user_data of the launched instances
know which explicit image to use.

I tested this by bypassing the filter on the CircleCI build workflow
configuration, but I can finally confirm that this actually deploys and
works as expected.