Re-structure Board component
The board component was causing problems because InputCells were being
created with `undefined`, due to the fact that the board is first
rendered with an empty `this.state.board`. I changed the logic such that
if a space is found in the board state while iterating through it, OR
the board state is empty, pass an empty string to be rendered into the
component.

I renamed getCellValues and made it a pure function, and the board state
is just passed into it whenever the board needs to be re-rendered.

Lastly, I updated/added tests to better verify that when the state of
the Board or the InputCell is changed, the rendered component reflects
that correctly.