Serve static assets from the React build at root
As per Charles's recommendation, we're going to start packaging the
optimized frontend build, directly into the backend docker image, and
then we're going to utilize CloudFront to ensure that we get good load
shedding.

Since we're serving the React bundle at root, we can remove the code
that we have for creating a route at root, as well as the test that was
asserting what root returned.