Make Docker build use a second stage
By running a second stage, we can take only what the output was of the
optimized build of the React app, and we don't need to have node
installed anymore, since that was only installed so that we could run
`npm install`. Now that the bundle has already been created, just move
everything to a second stage for the backend to be ran.

This ultimately creates a smaller Docker container, which means it is
also faster to load up.