Set no-cache for the main page
index.html has references to uniquely named resource IDs. The main.XXX.js
and main.XXX.css have a unique identifiers in them so that if they get
cached by CloudFront, there won't be a problem for the browser, because
the browser will have to request these uniquely named resources. Even if
these uniquely named resources get cached, and we do a new deployment,
the newly build index.html will have a reference to a new set of
resources, which will make the browser request the new resources. So
we'll never have to worry about the main.XXX.js/css files being
de-synced from the backend.

The problem that remains, is if a browser/client caches the index.html
file. Then when we deploy a new version of our application, the old
index.html page with the wrong main.js/css files will get loaded. We
can't have the same approach of uniquely naming our index.html, because
then people will have to try navigating to a strangely named index page.
The solution here is to make sure that when we send back a response for
the index.html, we specify to browsers to never cache this page, so that
every single time someone comes to visit the index.html, they have to
re-download it, and see if the main.js/css pages have been updated.