Update all tests to increase code coverage
All of the tests on the frontend have been updated to increase their
code coverage. More specifically, the Board component now has testing
available for its asyncronous fetch function.

The InputCell now has a better test that tests to see if an invalid
input is tried after a valid one.

The NoInputCell class was fleshed out a little to become more of an
idiomatic React component by giving it properties.