Apply multiple moves when a new move is made
Previously, we were only rendering a board with its latest move being
made. This was to make it easier to write tests, since I wasn't sure how
to do the equivalent of @Before from JUnit, which we'd need to reset the
state of all games, since each game is held in a global state variable.