Prevent reuse of the same block
This will probably have to be revisited, in the case that we allow
players to make valid but incorrect moves, they'll have to go back and
reuse the same cells. I think that in the future, what we'll have to do
is just prevent incorrect moves. This seems like the easiest thing to
do.