Change game.create to PUT
CloudFront will cache GET requests, and since we're being lazy and we're
using the same URL to create games in a stateful server, using the GET
method on game.create was being cached, and so CloudFront would never
forward these requests to the actual server, which means we weren't
actually creating new games if the same client IP was hitting the
game.create endpoint. As such, let's update everything to use PUT which
is more canonical in REST interfaces, for creating a new resource.

Something to think about, is that PUTs are supposed to be idempotent,
and game.create is definitely not idempotent. Maybe this should actually
be a POST endpoint, but for now PUT works just fine. We're creating a
stateful server anyways, so how important is it that we stick to
standard definitions of the REST definitions which are supposed to only
be for stateless applications.