Reformat parseBoard as an instance method
The `parseBoard` method of the `Board` class is now restructured as a
`class` method instead of a `static` one so we can pass properties and
state to it directly, instead of using parameters. Eslilnt will complain
about having too many parameters in a function call (default is up to 4)
so just binding the method to `this` and using it as a `class` method
cuts down on the need to pass every needed thing as a parameter.

The `Board` class has also been updated in preparation for passing all
of the needed properties to the `InputCell` class so we can update the
state of the board after it sends a message to the server about a new
user input.