Send user input to the server
The `InputCell` class now uses the `sendMoveToServer()` function to send
user input to the server after the user has tried to fill in one of the
cells with a validated input (the characters 1-9).

Technically, a blank character is a valid input, because we need to
update the state of the `InputCell` to be nothing after the user has
deleted an invalid character. However, we do not want to send blank
input to the server, so this is also taken into account.