Create tests for sending user input to the server
The `InputCell` component now has definitive tests to determine the
validity of the kind of responses we might get back from the server.
The server itself may respond with `INVALID` stating the user input was
incorrect for that cell; `INCOMPLETE` stating the user input was correct
for that cell, but the board still has empty cells; or `COMPLETE`
stating the user input was correct for that cell and the board has now
been filled with input to completion. The server may not respond at all
and as such this is handled too.

The `Board` component's tests now use a single recallable object to pass
to enzyme's `shallow` function so if any more tests need to be created
we don't have to make an entirely new object with all of the necessary
properties to pass to the board. The `InputCell` also reflects this
change, too.

Finally, `.eslintrc.yml` was also changed because forcing the sorting of
JS keys and variables messes with our ability to pass headers in a
canonical format.