Fix typo error for InputCell's POST method
The InputCell's `POST` method `sendMoveToServer` had a typo where
the field `header` should have been `header_s_`.
After this was addressed it was also apparent that the CORS preflight
`OPTIONS` request was being blocked by the server because the `OPTIONS`
check was not allowed for the `game.move` method in `game.py`. This was
also updated to fix the issue.

Now we should have no problems sending requests and responses from the
server to the client.