Avoid instantiating app in __init__
Since the web server was getting instantiated in __init__.py, we were
instantiating a new server every single time we imported the sudokurace
package. We only want to instantiate the app if we're going to be using
the app itself. Not instantiate it every single time we use something
from this package.