Change move POST format
The backend was changed to no longer nest the move information. There
was no reason for it to be nested in the first place, so I'm not sure
why I did it.

package-lock.json changes seem to have just come with the fact that I
did an npm install to test my change.