Increase max instance count
I'm wondering if the reason that we have downtime during a deploy, is
because the ASG is only allowed to have 1 instance at a time. So when we
bring up the new one for the deploy, the old one gets removed
immediately. I'm hoping that increasing the max count to 2 will allow
for the old and new instances to exist at once. I haven't tested this,
but it seems like a reasonable guess.