Use cheaper instance type
Since this project is going to get less attention, I've opted to switch
to using a T-nano which is smaller than the micro. It just so happens
that they came out with the 3rd gen T instances, which are even cheaper,
so we'll upgrade to using t3 instances as well.